added ResourceAccessService & finished auth

+++ Backend/src/main/java/com/project/backend/controller/JobController.java
@@ -2,10 +2,13 @@ package com.project.backend.controller;
 
 import com.project.backend.dto.JobDto;
 import com.project.backend.service.JobService;
+import com.project.backend.service.ResourceAccessService;
+import org.apache.tomcat.util.http.parser.Authorization;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.annotation.Secured;
+import org.springframework.security.core.Authentication;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -15,6 +18,8 @@ public class JobController {
 
     @Autowired
     JobService jobService;
+    @Autowired
+    ResourceAccessService resourceAccessService;
 
     @PostMapping("/job")
     public ResponseEntity<JobDto> addJob(@RequestBody JobDto jobDto) {
@@ -33,14 +38,18 @@ public class JobController {
     @DeleteMapping("/job")
     @Secured("ROLE_COMPANY")
     @ResponseBody
-    public ResponseEntity<JobDto> deleteJob(@RequestParam(value = "id") Integer jobId){
+    public ResponseEntity<JobDto> deleteJob(@RequestParam(value = "id") Integer jobId, Authentication authentication){
         JobDto jobDto = jobService.getJob(jobId);
 
         if (jobDto == null) {
             return new ResponseEntity<>(HttpStatus.NOT_FOUND);
         }
-        jobService.deleteJob(jobDto);
-
-        return ResponseEntity.ok(jobDto);
+        if(resourceAccessService.checkAccessJob(authentication, jobId)) {
+            jobService.deleteJob(jobDto);
+            return ResponseEntity.ok(jobDto);
+        }
+        else return ResponseEntity
+                .status(HttpStatus.FORBIDDEN)
+                .body(jobDto);
     }
 }
